macOS: do symlinks for files directly instead of folders (stow)
When a folder contains a single file or folder, stow does the
symlink on the folder not on the deeper file in the file tree
(the rule is probably more complex than that in reality).

This causes problem when this folder contains one or more
other files/folders in an extension of my dotfiles
(work-specific extension for example).
Stow will not be able to symlink the other files and will fail.

To fix this problem, we can force stow to create a symlink by file
instead of minimizing with folders symlinks.

The price to pay is not zero but not important: when creating a new
file in a folder containing previously one folder or file, we will
need to re-run the `make` command.

We can even argue that doing so is better because it simplifies the
times when a new `make` command is needed: every time we create or
remove a file.