macOS: rename brew-rebuild command
The `brew-rebuild` alias makes no sense.
It is based on NixOS's command `nixos-rebuild`
but in case of this brew command, it doesn't
rebuild anything.

This could have been renamed to `brew-update`
or `brew-up` but it feels a bit lame to me.

`brew-refill` seems nice and aligned with
the beer thematic of homebrew 🍻.